requirements.txt
altgraph==0.17.5
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
fuzzysearch==0.8.1
fuzzywuzzy==0.18.0
idna==3.11
igdb-api-v4==0.3.3
Levenshtein==0.27.3
linkify-it-py==2.0.3
macholib==1.16.4
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
packaging==25.0
platformdirs==4.5.1
protobuf==6.33.2
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
python-dotenv==1.2.1
python-Levenshtein==0.27.3
RapidFuzz==3.14.3
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
shiboken6==6.10.1
textual==6.10.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.6.2
Wikipedia-API==0.8.1
